Fix implicit conversion warning in timeout.c for newer Clang versions (#14811)
## Problem

Build fails on macOS with Xcode 26.4 beta (Clang 21):

```
timeout.c:157:21: error: implicit conversion from 'long long' to 'long double' changes value from 9223372036854775807 to 9223372036854775808 [-Werror,-Wimplicit-const-int-float-conversion]
  157 |         if (ftval > LLONG_MAX) {
      |                   ~ ^~~~~~~~~
/Applications/Xcode_26.4_beta.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/lib/clang/21/include/limits.h:109:20: note: expanded from macro 'LLONG_MAX'
  109 | #define LLONG_MAX  __LONG_LONG_MAX__
      |                    ^~~~~~~~~~~~~~~~~
<built-in>:64:27: note: expanded from macro '__LONG_LONG_MAX__'
   64 | #define __LONG_LONG_MAX__ 9223372036854775807LL
      |                           ^~~~~~~~~~~~~~~~~~~~~
1 error generated.
make[1]: *** [timeout.o] Error 1
make: *** [all] Error 2
```
## Cause

Comparing `long double` with `LLONG_MAX` causes implicit conversion with
precision loss. Clang 21 has stricter warnings for this, and `-Werror`
treats it as an error.

## Fix

Add explicit cast to make the conversion intentional.

---------

+++ src/timeout.c
@@ -153,7 +153,7 @@ int getTimeoutFromObjectOrReply(client *c, robj *object, mstime_t *timeout, int 
             return C_ERR;
 
         ftval *= 1000.0;  /* seconds => millisec */
-        if (ftval > LLONG_MAX) {
+        if (ftval > (long double)LLONG_MAX) {
             addReplyError(c, "timeout is out of range");
             return C_ERR;
         }
